Separate classes into multiple java files
